format

+++ test/ParaswapScaleHelper.t.sol
@@ -6,10 +6,11 @@ import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import "../contracts/router/swap-aggregator/ParaswapScaleHelper.sol";
 
 contract ParaswapScaleHelperTestHarness is ParaswapScaleHelper {
-    function exposedParaswapScaling(
-        bytes calldata rawCallData,
-        uint256 amountIn
-    ) external pure returns (bytes memory) {
+    function exposedParaswapScaling(bytes calldata rawCallData, uint256 amountIn)
+        external
+        pure
+        returns (bytes memory)
+    {
         return _paraswapScaling(rawCallData, amountIn);
     }
 }
@@ -22,14 +23,14 @@ contract ParaswapScaleHelperTest is Test {
     address constant TOKEN_DEST = address(0x6B175474E89094C44Da98b954EedeAC495271d0F);
     address constant EXECUTOR = address(0x6A000F20005980200259B80c5102003040001068);
     address constant BENEFICIARY = address(0x1);
-    
+
     uint256 constant ORIGINAL_AMOUNT = 1000;
     uint256 constant SCALED_AMOUNT = 2000; // 2x
     uint256 constant EXPECTED_TO_AMOUNT = 2000;
     uint256 constant EXPECTED_QUOTED_AMOUNT = 2100;
     uint256 constant EXPECTED_SCALED_TO_AMOUNT = 4000; // 2x
     uint256 constant EXPECTED_SCALED_QUOTED_AMOUNT = 4200; // 2x
-    
+
     bytes constant PERMIT = "01";
     bytes constant EXECUTOR_DATA = "02";
     uint256 constant PARTNER_AND_FEE = 123;
@@ -52,12 +53,7 @@ contract ParaswapScaleHelperTest is Test {
 
         // Create the original calldata
         bytes memory rawCallData = abi.encodeWithSelector(
-            IAugustusV6.swapExactAmountIn.selector,
-            EXECUTOR,
-            data,
-            PARTNER_AND_FEE,
-            PERMIT,
-            EXECUTOR_DATA
+            IAugustusV6.swapExactAmountIn.selector, EXECUTOR, data, PARTNER_AND_FEE, PERMIT, EXECUTOR_DATA
         );
 
         // Scale the amount by 2x
@@ -65,25 +61,26 @@ contract ParaswapScaleHelperTest is Test {
 
         // Decode the scaled calldata and verify the amounts
         bytes memory callDataWithoutSelector = new bytes(scaledCallData.length - 4);
-        for (uint i = 0; i < callDataWithoutSelector.length; i++) {
+        for (uint256 i = 0; i < callDataWithoutSelector.length; i++) {
             callDataWithoutSelector[i] = scaledCallData[i + 4];
         }
-        
+
         (
             address decodedExecutor,
             IAugustusV6.GenericData memory decodedData,
             uint256 decodedPartnerAndFee,
             bytes memory decodedPermit,
             bytes memory decodedExecutorData
-        ) = abi.decode(
-            callDataWithoutSelector,
-            (address, IAugustusV6.GenericData, uint256, bytes, bytes)
-        );
+        ) = abi.decode(callDataWithoutSelector, (address, IAugustusV6.GenericData, uint256, bytes, bytes));
 
         // Check that the amounts have been scaled correctly
         assertEq(decodedData.fromAmount, SCALED_AMOUNT, "From amount should be scaled to the new amount");
         assertEq(decodedData.toAmount, EXPECTED_SCALED_TO_AMOUNT, "To amount should be scaled proportionally (2x)");
-        assertEq(decodedData.quotedAmount, EXPECTED_SCALED_QUOTED_AMOUNT, "Quoted amount should be scaled proportionally (2x)");
+        assertEq(
+            decodedData.quotedAmount,
+            EXPECTED_SCALED_QUOTED_AMOUNT,
+            "Quoted amount should be scaled proportionally (2x)"
+        );
 
         // Check that the executor and other parameters stayed the same
         assertEq(decodedExecutor, EXECUTOR, "Executor address should not change");
@@ -96,7 +93,6 @@ contract ParaswapScaleHelperTest is Test {
         assertEq(address(decodedData.destToken), TOKEN_DEST, "Destination token should not change");
         assertEq(decodedData.beneficiary, payable(BENEFICIARY), "Beneficiary should not change");
         assertEq(decodedData.metadata, bytes32(0), "Metadata should not change");
-        
     }
 
     function test_swapExactAmountInOnUniswapV2() public {
@@ -114,11 +110,7 @@ contract ParaswapScaleHelperTest is Test {
 
         // Create the original calldata
         bytes memory rawCallData = abi.encodeWithSelector(
-            IAugustusV6.swapExactAmountInOnUniswapV2.selector,
-            EXECUTOR,
-            data,
-            PARTNER_AND_FEE,
-            PERMIT
+            IAugustusV6.swapExactAmountInOnUniswapV2.selector, EXECUTOR, data, PARTNER_AND_FEE, PERMIT
         );
 
         // Scale the amount by 2x
@@ -126,24 +118,25 @@ contract ParaswapScaleHelperTest is Test {
 
         // Decode the scaled calldata and verify the amounts
         bytes memory callDataWithoutSelector = new bytes(scaledCallData.length - 4);
-        for (uint i = 0; i < callDataWithoutSelector.length; i++) {
+        for (uint256 i = 0; i < callDataWithoutSelector.length; i++) {
             callDataWithoutSelector[i] = scaledCallData[i + 4];
         }
-        
+
         (
             address decodedExecutor,
             IAugustusV6.UniswapV2Data memory decodedData,
             uint256 decodedPartnerAndFee,
             bytes memory decodedPermit
-        ) = abi.decode(
-            callDataWithoutSelector,
-            (address, IAugustusV6.UniswapV2Data, uint256, bytes)
-        );
+        ) = abi.decode(callDataWithoutSelector, (address, IAugustusV6.UniswapV2Data, uint256, bytes));
 
         // Check that the amounts have been scaled correctly
         assertEq(decodedData.fromAmount, SCALED_AMOUNT, "From amount should be scaled to the new amount");
         assertEq(decodedData.toAmount, EXPECTED_SCALED_TO_AMOUNT, "To amount should be scaled proportionally (2x)");
-        assertEq(decodedData.quotedAmount, EXPECTED_SCALED_QUOTED_AMOUNT, "Quoted amount should be scaled proportionally (2x)");
+        assertEq(
+            decodedData.quotedAmount,
+            EXPECTED_SCALED_QUOTED_AMOUNT,
+            "Quoted amount should be scaled proportionally (2x)"
+        );
 
         // Check that the executor and other parameters stayed the same
         assertEq(decodedExecutor, EXECUTOR, "Executor address should not change");
@@ -173,11 +166,7 @@ contract ParaswapScaleHelperTest is Test {
 
         // Create the original calldata
         bytes memory rawCallData = abi.encodeWithSelector(
-            IAugustusV6.swapExactAmountInOnUniswapV3.selector,
-            EXECUTOR,
-            data,
-            PARTNER_AND_FEE,
-            PERMIT
+            IAugustusV6.swapExactAmountInOnUniswapV3.selector, EXECUTOR, data, PARTNER_AND_FEE, PERMIT
         );
 
         // Scale the amount by 2x
@@ -185,24 +174,25 @@ contract ParaswapScaleHelperTest is Test {
 
         // Decode the scaled calldata and verify the amounts
         bytes memory callDataWithoutSelector = new bytes(scaledCallData.length - 4);
-        for (uint i = 0; i < callDataWithoutSelector.length; i++) {
+        for (uint256 i = 0; i < callDataWithoutSelector.length; i++) {
             callDataWithoutSelector[i] = scaledCallData[i + 4];
         }
-        
+
         (
             address decodedExecutor,
             IAugustusV6.UniswapV3Data memory decodedData,
             uint256 decodedPartnerAndFee,
             bytes memory decodedPermit
-        ) = abi.decode(
-            callDataWithoutSelector,
-            (address, IAugustusV6.UniswapV3Data, uint256, bytes)
-        );
+        ) = abi.decode(callDataWithoutSelector, (address, IAugustusV6.UniswapV3Data, uint256, bytes));
 
         // Check that the amounts have been scaled correctly
         assertEq(decodedData.fromAmount, SCALED_AMOUNT, "From amount should be scaled to the new amount");
         assertEq(decodedData.toAmount, EXPECTED_SCALED_TO_AMOUNT, "To amount should be scaled proportionally (2x)");
-        assertEq(decodedData.quotedAmount, EXPECTED_SCALED_QUOTED_AMOUNT, "Quoted amount should be scaled proportionally (2x)");
+        assertEq(
+            decodedData.quotedAmount,
+            EXPECTED_SCALED_QUOTED_AMOUNT,
+            "Quoted amount should be scaled proportionally (2x)"
+        );
 
         // Check that the executor and other parameters stayed the same
         assertEq(decodedExecutor, EXECUTOR, "Executor address should not change");
@@ -232,12 +222,7 @@ contract ParaswapScaleHelperTest is Test {
 
         // Create the original calldata
         bytes memory rawCallData = abi.encodeWithSelector(
-            IAugustusV6.swapExactAmountInOnBalancerV2.selector,
-            EXECUTOR,
-            data,
-            PARTNER_AND_FEE,
-            PERMIT,
-            extraData
+            IAugustusV6.swapExactAmountInOnBalancerV2.selector, EXECUTOR, data, PARTNER_AND_FEE, PERMIT, extraData
         );
 
         // Scale the amount by 2x
@@ -245,25 +230,26 @@ contract ParaswapScaleHelperTest is Test {
 
         // Decode the scaled calldata and verify the amounts
         bytes memory callDataWithoutSelector = new bytes(scaledCallData.length - 4);
-        for (uint i = 0; i < callDataWithoutSelector.length; i++) {
+        for (uint256 i = 0; i < callDataWithoutSelector.length; i++) {
             callDataWithoutSelector[i] = scaledCallData[i + 4];
         }
-        
+
         (
             address decodedExecutor,
             IAugustusV6.BalancerV2Data memory decodedData,
             uint256 decodedPartnerAndFee,
             bytes memory decodedPermit,
             bytes memory decodedExtraData
-        ) = abi.decode(
-            callDataWithoutSelector,
-            (address, IAugustusV6.BalancerV2Data, uint256, bytes, bytes)
-        );
+        ) = abi.decode(callDataWithoutSelector, (address, IAugustusV6.BalancerV2Data, uint256, bytes, bytes));
 
         // Check that the amounts have been scaled correctly
         assertEq(decodedData.fromAmount, SCALED_AMOUNT, "From amount should be scaled to the new amount");
         assertEq(decodedData.toAmount, EXPECTED_SCALED_TO_AMOUNT, "To amount should be scaled proportionally (2x)");
-        assertEq(decodedData.quotedAmount, EXPECTED_SCALED_QUOTED_AMOUNT, "Quoted amount should be scaled proportionally (2x)");
+        assertEq(
+            decodedData.quotedAmount,
+            EXPECTED_SCALED_QUOTED_AMOUNT,
+            "Quoted amount should be scaled proportionally (2x)"
+        );
 
         // Check that the executor and other parameters stayed the same
         assertEq(decodedExecutor, EXECUTOR, "Executor address should not change");
@@ -292,11 +278,7 @@ contract ParaswapScaleHelperTest is Test {
 
         // Create the original calldata
         bytes memory rawCallData = abi.encodeWithSelector(
-            IAugustusV6.swapExactAmountInOnCurveV1.selector,
-            EXECUTOR,
-            data,
-            PARTNER_AND_FEE,
-            PERMIT
+            IAugustusV6.swapExactAmountInOnCurveV1.selector, EXECUTOR, data, PARTNER_AND_FEE, PERMIT
         );
 
         // Scale the amount by 2x
@@ -304,24 +286,25 @@ contract ParaswapScaleHelperTest is Test {
 
         // Decode the scaled calldata and verify the amounts
         bytes memory callDataWithoutSelector = new bytes(scaledCallData.length - 4);
-        for (uint i = 0; i < callDataWithoutSelector.length; i++) {
+        for (uint256 i = 0; i < callDataWithoutSelector.length; i++) {
             callDataWithoutSelector[i] = scaledCallData[i + 4];
         }
-        
+
         (
             address decodedExecutor,
             IAugustusV6.CurveV1Data memory decodedData,
             uint256 decodedPartnerAndFee,
             bytes memory decodedPermit
-        ) = abi.decode(
-            callDataWithoutSelector,
-            (address, IAugustusV6.CurveV1Data, uint256, bytes)
-        );
+        ) = abi.decode(callDataWithoutSelector, (address, IAugustusV6.CurveV1Data, uint256, bytes));
 
         // Check that the amounts have been scaled correctly
         assertEq(decodedData.fromAmount, SCALED_AMOUNT, "From amount should be scaled to the new amount");
         assertEq(decodedData.toAmount, EXPECTED_SCALED_TO_AMOUNT, "To amount should be scaled proportionally (2x)");
-        assertEq(decodedData.quotedAmount, EXPECTED_SCALED_QUOTED_AMOUNT, "Quoted amount should be scaled proportionally (2x)");
+        assertEq(
+            decodedData.quotedAmount,
+            EXPECTED_SCALED_QUOTED_AMOUNT,
+            "Quoted amount should be scaled proportionally (2x)"
+        );
 
         // Check that the executor and other parameters stayed the same
         assertEq(decodedExecutor, EXECUTOR, "Executor address should not change");
@@ -355,11 +338,7 @@ contract ParaswapScaleHelperTest is Test {
 
         // Create the original calldata
         bytes memory rawCallData = abi.encodeWithSelector(
-            IAugustusV6.swapExactAmountInOnCurveV2.selector,
-            EXECUTOR,
-            data,
-            PARTNER_AND_FEE,
-            PERMIT
+            IAugustusV6.swapExactAmountInOnCurveV2.selector, EXECUTOR, data, PARTNER_AND_FEE, PERMIT
         );
 
         // Scale the amount by 2x
@@ -367,24 +346,25 @@ contract ParaswapScaleHelperTest is Test {
 
         // Decode the scaled calldata and verify the amounts
         bytes memory callDataWithoutSelector = new bytes(scaledCallData.length - 4);
-        for (uint i = 0; i < callDataWithoutSelector.length; i++) {
+        for (uint256 i = 0; i < callDataWithoutSelector.length; i++) {
             callDataWithoutSelector[i] = scaledCallData[i + 4];
         }
-        
+
         (
             address decodedExecutor,
             IAugustusV6.CurveV2Data memory decodedData,
             uint256 decodedPartnerAndFee,
             bytes memory decodedPermit
-        ) = abi.decode(
-            callDataWithoutSelector,
-            (address, IAugustusV6.CurveV2Data, uint256, bytes)
-        );
+        ) = abi.decode(callDataWithoutSelector, (address, IAugustusV6.CurveV2Data, uint256, bytes));
 
         // Check that the amounts have been scaled correctly
         assertEq(decodedData.fromAmount, SCALED_AMOUNT, "From amount should be scaled to the new amount");
         assertEq(decodedData.toAmount, EXPECTED_SCALED_TO_AMOUNT, "To amount should be scaled proportionally (2x)");
-        assertEq(decodedData.quotedAmount, EXPECTED_SCALED_QUOTED_AMOUNT, "Quoted amount should be scaled proportionally (2x)");
+        assertEq(
+            decodedData.quotedAmount,
+            EXPECTED_SCALED_QUOTED_AMOUNT,
+            "Quoted amount should be scaled proportionally (2x)"
+        );
 
         // Check that the executor and other parameters stayed the same
         assertEq(decodedExecutor, EXECUTOR, "Executor address should not change");
@@ -397,7 +377,11 @@ contract ParaswapScaleHelperTest is Test {
         assertEq(decodedData.curveData, 12345, "Curve data should not change");
         assertEq(decodedData.i, 0, "i value should not change");
         assertEq(decodedData.j, 1, "j value should not change");
-        assertEq(decodedData.poolAddress, address(0x1234567890123456789012345678901234567890), "Pool address should not change");
+        assertEq(
+            decodedData.poolAddress,
+            address(0x1234567890123456789012345678901234567890),
+            "Pool address should not change"
+        );
         assertEq(decodedData.beneficiary, payable(BENEFICIARY), "Beneficiary should not change");
         assertEq(decodedData.metadata, bytes32(0), "Metadata should not change");
     }
@@ -435,32 +419,24 @@ contract ParaswapScaleHelperTest is Test {
         });
 
         // Create the original calldata
-        bytes memory rawCallData = abi.encodeWithSelector(
-            IAugustusV6.swapOnAugustusRFQTryBatchFill.selector,
-            EXECUTOR,
-            data,
-            orders,
-            PERMIT
-        );
+        bytes memory rawCallData =
+            abi.encodeWithSelector(IAugustusV6.swapOnAugustusRFQTryBatchFill.selector, EXECUTOR, data, orders, PERMIT);
 
         // Scale the amount by 2x
         bytes memory scaledCallData = helper.exposedParaswapScaling(rawCallData, SCALED_AMOUNT);
 
         // Decode the scaled calldata and verify the amounts
         bytes memory callDataWithoutSelector = new bytes(scaledCallData.length - 4);
-        for (uint i = 0; i < callDataWithoutSelector.length; i++) {
+        for (uint256 i = 0; i < callDataWithoutSelector.length; i++) {
             callDataWithoutSelector[i] = scaledCallData[i + 4];
         }
-        
+
         (
             address decodedExecutor,
             IAugustusV6.AugustusRFQData memory decodedData,
             IAugustusV6.OrderInfo[] memory decodedOrders,
             bytes memory decodedPermit
-        ) = abi.decode(
-            callDataWithoutSelector,
-            (address, IAugustusV6.AugustusRFQData, IAugustusV6.OrderInfo[], bytes)
-        );
+        ) = abi.decode(callDataWithoutSelector, (address, IAugustusV6.AugustusRFQData, IAugustusV6.OrderInfo[], bytes));
 
         // Check that the amounts have been scaled correctly
         assertEq(decodedData.fromAmount, SCALED_AMOUNT, "From amount should be scaled to the new amount");
@@ -474,7 +450,7 @@ contract ParaswapScaleHelperTest is Test {
         assertEq(decodedData.wrapApproveDirection, 1, "Wrap direction should not change");
         assertEq(decodedData.beneficiary, payable(BENEFICIARY), "Beneficiary should not change");
         assertEq(decodedData.metadata, bytes32(0), "Metadata should not change");
-        
+
         // Check that orders data is preserved
         assertEq(decodedOrders.length, 1, "Should have 1 order");
         assertEq(decodedOrders[0].order.nonceAndMeta, 12345, "Order nonce should not change");
@@ -497,30 +473,20 @@ contract ParaswapScaleHelperTest is Test {
         });
 
         // Create the original calldata
-        bytes memory rawCallData = abi.encodeWithSelector(
-            IAugustusV6.swapExactAmountInOutOnMakerPSM.selector,
-            EXECUTOR,
-            data,
-            PERMIT
-        );
+        bytes memory rawCallData =
+            abi.encodeWithSelector(IAugustusV6.swapExactAmountInOutOnMakerPSM.selector, EXECUTOR, data, PERMIT);
 
         // Scale the amount by 2x
         bytes memory scaledCallData = helper.exposedParaswapScaling(rawCallData, SCALED_AMOUNT);
 
         // Decode the scaled calldata and verify the amounts
         bytes memory callDataWithoutSelector = new bytes(scaledCallData.length - 4);
-        for (uint i = 0; i < callDataWithoutSelector.length; i++) {
+        for (uint256 i = 0; i < callDataWithoutSelector.length; i++) {
             callDataWithoutSelector[i] = scaledCallData[i + 4];
         }
-        
-        (
-            address decodedExecutor,
-            IAugustusV6.MakerPSMData memory decodedData,
-            bytes memory decodedPermit
-        ) = abi.decode(
-            callDataWithoutSelector,
-            (address, IAugustusV6.MakerPSMData, bytes)
-        );
+
+        (address decodedExecutor, IAugustusV6.MakerPSMData memory decodedData, bytes memory decodedPermit) =
+            abi.decode(callDataWithoutSelector, (address, IAugustusV6.MakerPSMData, bytes));
 
         // Check that the amounts have been scaled correctly
         assertEq(decodedData.fromAmount, SCALED_AMOUNT, "From amount should be scaled to the new amount");
@@ -543,14 +509,10 @@ contract ParaswapScaleHelperTest is Test {
 
     function test_InvalidSelector() public {
         // Test that the scaling function reverts for an invalid selector
-        bytes memory invalidCallData = abi.encodeWithSelector(
-            bytes4(keccak256("invalidFunction()")),
-            address(0x1),
-            100,
-            200
-        );
-        
+        bytes memory invalidCallData =
+            abi.encodeWithSelector(bytes4(keccak256("invalidFunction()")), address(0x1), 100, 200);
+
         vm.expectRevert("ParaswapScaleHelper: Unsupported swap selector");
         helper.exposedParaswapScaling(invalidCallData, 1000);
     }
-} 
+}
